requirements.txt
chardet==5.2.0
pillow==10.4.0
PyPDF2==3.0.1
reportlab==4.2.2
